// event_user_frontend/src/types/eventType.ts
export type eventType = {
  id?: number;
  title?: string;
  description?: string;
  startDate?: string;
  startTime?: string;
  endTime?: string;
  endDate?: string;
  localization?: string;
  location?: string;
  lotation?: number | string;
  capacity?: number | string;
  image?: string;
  isActive?: boolean;
  updatedAt?: string;
  deletedAt?: string;
};